Some signed-unsigned comparisons fixed

+++ neo/d3xp/AF.cpp
@@ -892,7 +892,7 @@ bool idAF::Load( idEntity *ent, const char *fileName ) {
 		idAFConstraint *constraint = physicsObj.GetConstraint( i );
 		for ( j = 0; j < file->constraints.Num(); j++ ) {
 			if ( file->constraints[j]->name.Icmp( constraint->GetName() ) == 0 &&
-					file->constraints[j]->type == constraint->GetType() ) {
+					(int)file->constraints[j]->type == (int)constraint->GetType() ) {
 				break;
 			}
 		}
